Rename Slender.ServiceReigstrations to Slender.Dependencies
#18
