+++ Dependencies.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Slender.ServiceRegistrations", "src\Slender.ServiceRegistrations\Slender.ServiceRegistrations.csproj", "{E6A17F75-9572-42CF-A629-92EEF63FCCD8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Slender.Dependencies", "src\Slender.Dependencies\Slender.Dependencies.csproj", "{E6A17F75-9572-42CF-A629-92EEF63FCCD8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Slender.ServiceRegistrations.Tests", "test\Slender.ServiceRegistrations.Tests\Slender.ServiceRegistrations.Tests.csproj", "{9E851DD9-F7EB-4286-809C-7C757DD6059B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Slender.Dependencies.Tests", "test\Slender.Dependencies.Tests\Slender.Dependencies.Tests.csproj", "{9E851DD9-F7EB-4286-809C-7C757DD6059B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
