+++ src/Slender.Dependencies/Dependency.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Slender.ServiceRegistrations
+namespace Slender.Dependencies
 {
 
     /// <summary>
