+++ src/Slender.Dependencies/DependencyFactory.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Slender.ServiceRegistrations
+namespace Slender.Dependencies
 {
 
     /// <summary>
