+++ src/Slender.Dependencies/IDependencyBehaviour.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Slender.ServiceRegistrations
+namespace Slender.Dependencies
 {
 
     /// <summary>
