+++ test/Slender.Dependencies.Tests/Unit/DependencyCollectionTests.Transient.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using Xunit;
 
-namespace Slender.ServiceRegistrations.Tests.Unit
+namespace Slender.Dependencies.Tests.Unit
 {
 
     public partial class DependencyCollectionTests
